Move connection stuff to its own unit
