+++ tests/api/test_connections.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 import pytest
-from snowflake.cli.api.cli_global_context import _ConnectionContext
+from snowflake.cli.api.connections import ConnectionContext
 
 
 @pytest.mark.parametrize(
@@ -21,16 +21,16 @@ from snowflake.cli.api.cli_global_context import _ConnectionContext
     [
         (
             {"connection_name": "myconn"},
-            "_ConnectionContext(_connection_name = 'myconn', _enable_diag = False, _temporary_connection = False)",
+            "ConnectionContext(_connection_name = 'myconn', _enable_diag = False, _temporary_connection = False)",
         ),
         (
             {"temporary_connection": True, "account": "myacct", "user": "myuser"},
-            "_ConnectionContext(_account = 'myacct', _user = 'myuser', _enable_diag = False, _temporary_connection = True)",
+            "ConnectionContext(_account = 'myacct', _user = 'myuser', _enable_diag = False, _temporary_connection = True)",
         ),
     ],
 )
 def test_connection_context_repr(args: dict, expected: str):
-    ctx = _ConnectionContext()
+    ctx = ConnectionContext()
     for k, v in args.items():
         getattr(ctx, f"set_{k}")(v)
     assert repr(ctx) == expected
